fix: move home page folder from components to features/homePage
