+++ src/features/homePage/components/PollTimeLength.tsx
@@ -1,9 +1,9 @@
 import type { Dispatch, FC } from "react"
-import { StyledLabel, StyledSelect } from "../styledHTMLElements/FloatingStyles"
 import { createAndIncrementFill } from "~/utils/helpers"
 import React from "react"
-import { type PollLength } from "./types"
 import { type PollLengthAction } from "~/reducers/pollLengthReducer"
+import { type PollLength } from "../types"
+import { StyledLabel, StyledSelect } from "~/components/styledHTMLElements/FloatingStyles"
 
 const PollLengthOptions: FC<{ length: number; minLength?: number }> = ({
 	length,
