+++ src/features/homePage/index.ts
@@ -1,2 +1,3 @@
 export * from "./components"
 export * from "./hooks"
+export * from "./types"
